Added option of extracting day and night values separately.  Pushed through all downstream code.
Errors fixed in gridding_cam.py to apply n_obs from daily to monthly and also in the way the masks were being unset at the 3hrly gridding stage in utils.py.

Have reduced the completeness requirements down to 3 obs in a day and at least 1 obs in a pentad.